+++ extensions/Worker.Extensions.CosmosDB/src/CosmosDBInputAttribute.cs
@@ -2,20 +2,44 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
+using System.Text.Json.Serialization;
 using Microsoft.Azure.Functions.Worker.Converters;
 using Microsoft.Azure.Functions.Worker.Extensions.Abstractions;
 
 namespace Microsoft.Azure.Functions.Worker
 {
-    [AllowConverterFallback(false)]
     [InputConverter(typeof(CosmosDBConverter))]
+    [ConverterFallbackBehavior(ConverterFallbackBehavior.Default)]
     public sealed class CosmosDBInputAttribute : InputBindingAttribute
     {
         /// <summary>
         /// Constructs a new instance.
+        /// Use this constructor when binding to a CosmosClient.
+        /// </summary>
+        public CosmosDBInputAttribute()
+        {
+            DatabaseName = string.Empty;
+            ContainerName = string.Empty;
+        }
+
+        /// <summary>
+        /// Constructs a new instance with the specified database name.
+        /// Use this constructor when binding to a Database.
+        /// </summary>
+        /// <param name="databaseName">The CosmosDB database name.</param>
+        public CosmosDBInputAttribute(string databaseName)
+        {
+            DatabaseName = databaseName;
+            ContainerName = string.Empty;
+        }
+
+        /// <summary>
+        /// Constructs a new instance with the specified database and container names.
+        /// Use this constructor when binding to a Container or a POCO.
         /// </summary>
         /// <param name="databaseName">The CosmosDB database name.</param>
         /// <param name="containerName">The CosmosDB container name.</param>
+        [JsonConstructor]
         public CosmosDBInputAttribute(string databaseName, string containerName)
         {
             DatabaseName = databaseName;